feat(containers): image deployment caching
Adds `Edgehog.Containers.Image.Deployment` resource representing the
current state of the image on a device

`Image.Deployment` works like the `Deployment` resource, representing
the `many_to_many` relationship between an image and a device. It keeps
track of the current state of an image on the device and gets updated
when the device publishes on the `AvailableImages` interface.

Signed-off-by: Luca Zaninotto <luca.zaninotto@secomind.com>