feat!(containers): volume deployment caching
Adds `Edgehog.Containers.Volume.Deployment` resource representing the
current state of a volume on a device.

`Volume.Deployment` works like the `Deployment` resource, representing
the `many_to_many` relationship between an volume and a device. It keeps
track of the current state of an volume on the device and gets updated
when the device publishes on the `AvailableVolumes` interface.

Adds a previously skipped device status `:created_volumes` in `DeploymentStatus`

Signed-off-by: Luca Zaninotto <luca.zaninotto@secomind.com>